feat(flow-editor): refine EitherFlowAdvancedExample with enhanced UI and CRUD capabilities
